repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/14/LLMatic/group.py
class Group:
	def __init__(self, group_name, admin):
		self.group_name = group_name
		self.admin = admin
		self.participants = {admin: 'admin'}
		self.group_chat = []

	def add_participant(self, participant):
		if participant not in self.participants:
			self.participants[participant] = 'participant'

	def remove_participant(self, participant):
		if participant in self.participants and self.participants[participant] != 'admin':
			del self.participants[participant]

	def assign_admin(self, participant):
		if participant in self.participants:
			self.participants[participant] = 'admin'

	def add_message(self, message):
		self.group_chat.append(message)
